migrated from livedata to kotlin coroutine and flow

+++ app/src/main/java/dreammaker/android/expensetracker/ui/history/viewhistory/ViewHistoryItemFragment.kt
@@ -118,7 +118,7 @@ class ViewHistoryItemFragment: Fragment(), MenuProvider {
 
     }
 
-    private fun prepareAmount(amount: Float) {
+    private fun prepareAmount(amount: Double) {
         binding.amount.text = amount.toCurrencyString()
     }
